修改表结构，字段类型

+++ project-common/src/main/java/com/software/utils/SecurityUtils.java
@@ -51,9 +51,9 @@ public class SecurityUtils {
      *
      * @return 当前登录用户id
      */
-    public static Long getCurrentUserId() {
+    public static Integer getCurrentUserId() {
         UserDetails userDetails = getCurrentUser();
-        return new JSONObject(new JSONObject(userDetails).get("user")).get("id", Long.class);
+        return new JSONObject(new JSONObject(userDetails).get("user")).get("id", Integer.class);
     }
 
     /**
