+++ project-system/src/main/java/com/software/system/service/impl/UserRoleServiceImpl.java
@@ -17,7 +17,7 @@ import java.util.Arrays;
 public class UserRoleServiceImpl extends ServiceImpl<UserRoleMapper, UserRole> implements UserRoleService {
 
     @Override
-    public void deleteUserRoleByUserId(Long[] userIds) {
+    public void deleteUserRoleByUserId(Integer[] userIds) {
         if (userIds.length > 0) {
             this.baseMapper.delete(new LambdaQueryWrapper<UserRole>().in(UserRole::getUserId, Arrays.asList(userIds)));
         }
